dhrishyamitra/drishyamitra-backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
Werkzeug==3.1.3

opencv-python==4.11.0.86
numpy==2.1.3
DeepFace==0.0.93

groq==0.18.0
requests==2.32.3
